64-bit fix. Patch by Andrzej Dopierała.
2004-10-11  Christian Persch  <chpe@cvs.gnome.org>

	* extensions/error-viewer/sgml-validator.c: (convert_to_utf8):

	64-bit fix. Patch by Andrzej Dopierała.

+++ extensions/error-viewer/sgml-validator.c
@@ -271,8 +271,8 @@ convert_to_utf8 (const char *file,
 	char *encoding;
 	char *base;
 	char *buf;
-	int buf_size = 4096;
-	int len;
+	gsize buf_size = 4096;
+	gsize len;
 	GIOChannel *in;
 	GIOChannel *out;
 	GIOStatus status;
